Move new run rules function into rules engine
